pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secure-zeph"
version = "0.1.0"
description = "Prompt-injection security middleware & evaluation platform for agentic systems"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "duckdb>=1.2.0",
    "pyyaml>=6.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]
day2 = [
    "anthropic>=0.40.0",
    "azure-ai-inference>=1.0.0",
    "azure-identity>=1.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["apps/gateway/tests", "apps/evalrunner/tests", "packages/core"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "S", "B"]
ignore = ["S101"]
